pyproject.toml
[tool.poetry]
name = "griptape"
version = "0.11.4"
description = "Modular Python framework for LLM workflows, tools, memory, and data."
authors = ["Griptape <hello@griptape.ai>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/griptape-ai/griptape"

packages = [
    {include = "griptape"}
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = ">=0.21"
openai = ">=0.27"
cohere = ">=4"
attrs = ">=22"
jinja2 = ">=3.1"
jsonschema = ">=4"
marshmallow = ">=3"
marshmallow-enum = ">=1.5"
graphlib = "*"
tiktoken = ">=0.3"
rich = ">=13"
stopit = "*"
transformers = ">=4"
huggingface-hub = ">=0.13"
schema = ">=0.7"
pyyaml = ">=6"
fastapi = ">=0.80"
uvicorn = ">= 0.20"
python-decouple=">=3"
boto3 = "^1.26.123"
tenacity = ">=8.0"
numpy = ">=1"
pinecone-client = ">=2"
PyPDF2 = ">=3"
trafilatura = ">= 1.6"
anthropic = "^0.2.10"
sqlalchemy = "~=1.0"
sqlalchemy-redshift = "*"
sentencepiece = "*"

[tool.poetry.group.test.dependencies]
pytest = "~=7.1"
pytest-mock = "*"
pytest-cover = "*"
twine = ">=4"
moto = {extras = ["dynamodb"], version = "^4.1.8"}


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.3"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
